Hook up tests; including some example tests for user methods
Here are a couple example tests for our project. There are a couple
things to note here. We are using a gem called factory_girl to create
fake test objects. Because there is no support for in-memory objects (as
there is with ActiveRecord), running the test suite will create lots of
clutter in your zombie_game_playground_test database. So, in
spec/spec_helper.rb (which is run at the beginning of the test suite) we
forcefully delete all things that are hanging around the database.

Anyway, so the structure of the spec/ directory should follow the
structure of app/. Filenames are important here -- i.e. tests for the
User model (app/models/user.rb) should be located in
spec/models/user_spec.rb. In the future we will probably want to break
out some of these tests into a spec/lib/zombie_game/ directory to test
the engine specifically.