Better use MongoMapper for cancan access queries
user.players.find(...) executes a query and returns an array or nil, in
the latter case nil has no .any? method, which causes an error.

user.players.where(...).exists? executes a query checking if the count
of the records is 1, rather than returning all results to Ruby and then
using Ruby's array methods to determine existence.