Add a player's game state (Human, Zombie, Deceased)
This adds a method to get the current state of the player. The way it
works is by beginning each player in the :human state and then
successively calling a chain of methods which can change the ":human"
state to something else if they want, based on whatever logic is
appropriate. Currently, this looks something like:
  Begin  -- :human
  Step 1 -- Has the player been tagged? Then change :human -> :zombie
  Step 2 -- Is the player an Original Zombie? Then change :human ->
            :zombie
  Step 3 -- Has the player been tagged, been fed, or been converted into
            a zombie in the past 48 hours? If not, change :zombie ->
            :deceased